// src/app/models/periciandos.ts
export class Periciandos {
    codigo: string;
    usuario: string;
    envolvimento: number;
    tipificacao: string;
    nao_identificado: string ='false';
    recem_nascido: string ='false';
    mae_nao_identificada: string ='false';
    tipo: string;
    cpf: string;
    rg: string;
    orgaoExp: number;
    passaporte: string;
    nome: string;
    nome_social: string;
    filiacao: string;
    filiacao1:string;
    filiacao2:string;
    data_nacimento: string;
    sexo: string;
    cutis: string;
    nacionalidade: string;
    naturalidade: string;
    deficiencia: string;
    idade: string ='false';
    fisica: string ='false';
    mental: string='false';
    auditiva: string='false';
    visual: string ='false';
    estado_civil: string;
    profissao: string;
    reliagao: string;
    graude_instrucao: string;
    orientacao_sexual: string;
    email: string;
    telefone: string;
    pais: string;
    uf: number = 29;
    cep: string;
    municipio: number;
    bairro: number;
    tipo_logradouro: number;
    complemento: string;
    data: string;
    hora: string;
    local_provavel: number;
    causas_provavel: string;
    uf2: string;
    cep2: string;
    municipio2: string;
    bairro2: string;
    tipo_logradouro2: string;
    complemento2: string;
    mesmo_endereco_pericando: string= 'false';
    mesmo_endereco_localfato: string ='false';
    exame_externo: string='false';
    tipo_local: number;
    descricao_local: string;
    estado_periciando: string;
    dataexamen1: string;
    horaexamen1: string;
    dataexamen2: string;
    horaexamen2: string;
    acidente_trabalho: string='false';
    quesitacoes: string;
    numguia: string;
    numocur: string;
    rascunho: string;
    tipo_guia: string;
    
    matricula: string='';
    descricao: string;
    tentativaHomicidio:string;
    homicidio:string;
    dataObito:string;

    //VALUES TO SHOW THE INFORMATION IN THE DOCUMENT TO PRINT 
    descricaoPrint: string='';
    informacoesComplementares: string = '';
}